[TNT-115] move: 연결 화면 trainer trainee 분리
